Day09

+++ app/src/main/java/com/sandervanderlinden/adventofcode2023/common/LineSumPuzzleSolver.java
@@ -2,7 +2,7 @@ package com.sandervanderlinden.adventofcode2023.common;
 
 public abstract class LineSumPuzzleSolver implements BasePuzzleSolver {
 
-    protected int sum;
+    protected long sum;
 
     @Override
     public void initializeSolver() {
@@ -11,7 +11,7 @@ public abstract class LineSumPuzzleSolver implements BasePuzzleSolver {
 
     @Override
     public void processLine(String line) {
-        int value = extractValueFromLine(line);
+        long value = extractValueFromLine(line);
         sum += value;
     }
 
@@ -27,5 +27,5 @@ public abstract class LineSumPuzzleSolver implements BasePuzzleSolver {
      * @param line The line of text from which destination extract the value.
      * @return The extracted value as an integer.
      */
-    protected abstract int extractValueFromLine(String line);
+    protected abstract long extractValueFromLine(String line);
 }
